didact.cpp: split out of Eseelte 100 computer in its own driver for clarity
